+++ packages/vite/src/node/optimizer/index.ts
@@ -31,12 +31,15 @@ const isDebugEnabled = _debug('vite:deps').enabled
 const jsExtensionRE = /\.js$/i
 const jsMapExtensionRE = /\.js\.map$/i
 
-export type ExportsData = ReturnType<typeof parse> & {
+export type ExportsData = {
+  hasImports: boolean
+  exports: readonly string[]
+  facade: boolean
   // es-module-lexer has a facade detection but isn't always accurate for our
   // use case when the module has default export
-  hasReExports?: true
+  hasReExports?: boolean
   // hint if the dep requires loading as jsx
-  jsxLoader?: true
+  jsxLoader?: boolean
 }
 
 export interface DepsOptimizer {
@@ -62,6 +65,13 @@ export interface DepOptimizationOptions {
    * vite project root. This will overwrite default entries inference.
    */
   entries?: string | string[]
+  /**
+   * Enable esbuild based scan phase, to get back to the optimized deps discovery
+   * strategy used in Vite v2
+   * @default false
+   * @experimental
+   */
+  devScan?: boolean
   /**
    * Force optimize listed dependencies (must be resolvable import paths,
    * cannot be globs).
@@ -475,7 +485,7 @@ export async function runOptimizeDeps(
     splitting: true,
     sourcemap: true,
     outdir: processingCacheDir,
-    ignoreAnnotations: true,
+    ignoreAnnotations: resolvedConfig.command !== 'build',
     metafile: true,
     define,
     plugins: [
@@ -585,33 +595,51 @@ export function newDepOptimizationProcessing(): DepOptimizationProcessing {
 // Convert to { id: src }
 export function depsFromOptimizedDepInfo(
   depsInfo: Record<string, OptimizedDepInfo>
-) {
+): Record<string, string> {
   return Object.fromEntries(
     Object.entries(depsInfo).map((d) => [d[0], d[1].src!])
   )
 }
 
-export function getOptimizedDepPath(id: string, config: ResolvedConfig) {
+export function getOptimizedDepPath(
+  id: string,
+  config: ResolvedConfig
+): string {
   return normalizePath(
     path.resolve(getDepsCacheDir(config), flattenId(id) + '.js')
   )
 }
 
-export function getDepsCacheDir(config: ResolvedConfig) {
-  const dirName = config.command === 'build' ? 'depsBuild' : 'deps'
+function getDepsCacheSuffix(config: ResolvedConfig): string {
+  let suffix = ''
+  if (config.command === 'build') {
+    suffix += '_build'
+    if (config.build.ssr) {
+      suffix += '_ssr'
+    }
+  }
+  return suffix
+}
+export function getDepsCacheDir(config: ResolvedConfig): string {
+  const dirName = 'deps' + getDepsCacheSuffix(config)
   return normalizePath(path.resolve(config.cacheDir, dirName))
 }
 
 function getProcessingDepsCacheDir(config: ResolvedConfig) {
-  const dirName = config.command === 'build' ? 'processingBuild' : 'processing'
+  const dirName = 'deps' + getDepsCacheSuffix(config) + '_temp'
   return normalizePath(path.resolve(config.cacheDir, dirName))
 }
 
-export function isOptimizedDepFile(id: string, config: ResolvedConfig) {
+export function isOptimizedDepFile(
+  id: string,
+  config: ResolvedConfig
+): boolean {
   return id.startsWith(getDepsCacheDir(config))
 }
 
-export function createIsOptimizedDepUrl(config: ResolvedConfig) {
+export function createIsOptimizedDepUrl(
+  config: ResolvedConfig
+): (url: string) => boolean {
   const { root } = config
   const depsCacheDir = getDepsCacheDir(config)
 
@@ -740,7 +768,6 @@ export async function extractExportsData(
   config: ResolvedConfig
 ): Promise<ExportsData> {
   await init
-  let exportsData: ExportsData
 
   const esbuildOptions = config.optimizeDeps?.esbuildOptions ?? {}
   if (config.optimizeDeps.extensions?.some((ext) => filePath.endsWith(ext))) {
@@ -753,34 +780,48 @@ export async function extractExportsData(
       write: false,
       format: 'esm'
     })
-    exportsData = parse(result.outputFiles[0].text) as ExportsData
-  } else {
-    const entryContent = fs.readFileSync(filePath, 'utf-8')
-    try {
-      exportsData = parse(entryContent) as ExportsData
-    } catch {
-      const loader = esbuildOptions.loader?.[path.extname(filePath)] || 'jsx'
-      debug(
-        `Unable to parse: ${filePath}.\n Trying again with a ${loader} transform.`
-      )
-      const transformed = await transformWithEsbuild(entryContent, filePath, {
-        loader
-      })
-      // Ensure that optimization won't fail by defaulting '.js' to the JSX parser.
-      // This is useful for packages such as Gatsby.
-      esbuildOptions.loader = {
-        '.js': 'jsx',
-        ...esbuildOptions.loader
-      }
-      exportsData = parse(transformed.code) as ExportsData
-      exportsData.jsxLoader = true
+    const [imports, exports, facade] = parse(result.outputFiles[0].text)
+    return {
+      hasImports: imports.length > 0,
+      exports,
+      facade
     }
-    for (const { ss, se } of exportsData[0]) {
+  }
+
+  let parseResult: ReturnType<typeof parse>
+  let usedJsxLoader = false
+
+  const entryContent = fs.readFileSync(filePath, 'utf-8')
+  try {
+    parseResult = parse(entryContent)
+  } catch {
+    const loader = esbuildOptions.loader?.[path.extname(filePath)] || 'jsx'
+    debug(
+      `Unable to parse: ${filePath}.\n Trying again with a ${loader} transform.`
+    )
+    const transformed = await transformWithEsbuild(entryContent, filePath, {
+      loader
+    })
+    // Ensure that optimization won't fail by defaulting '.js' to the JSX parser.
+    // This is useful for packages such as Gatsby.
+    esbuildOptions.loader = {
+      '.js': 'jsx',
+      ...esbuildOptions.loader
+    }
+    parseResult = parse(transformed.code)
+    usedJsxLoader = true
+  }
+
+  const [imports, exports, facade] = parseResult
+  const exportsData: ExportsData = {
+    hasImports: imports.length > 0,
+    exports,
+    facade,
+    hasReExports: imports.some(({ ss, se }) => {
       const exp = entryContent.slice(ss, se)
-      if (/export\s+\*\s+from/.test(exp)) {
-        exportsData.hasReExports = true
-      }
-    }
+      return /export\s+\*\s+from/.test(exp)
+    }),
+    jsxLoader: usedJsxLoader
   }
   return exportsData
 }
@@ -802,9 +843,9 @@ function needsInterop(
   ) {
     return true
   }
-  const [imports, exports] = exportsData
+  const { hasImports, exports } = exportsData
   // entry has no ESM syntax - likely CJS or UMD
-  if (!exports.length && !imports.length) {
+  if (!exports.length && !hasImports) {
     return true
   }