+++ packages/vite/src/node/optimizer/optimizer.ts
@@ -34,7 +34,9 @@ const debounceMs = 100
 
 const depsOptimizerMap = new WeakMap<ResolvedConfig, DepsOptimizer>()
 
-export function getDepsOptimizer(config: ResolvedConfig) {
+export function getDepsOptimizer(
+  config: ResolvedConfig
+): DepsOptimizer | undefined {
   // Workers compilation shares the DepsOptimizer from the main build
   return depsOptimizerMap.get(config.mainConfig || config)
 }
@@ -45,6 +47,8 @@ export async function initDepsOptimizer(
 ): Promise<DepsOptimizer> {
   const { logger } = config
   const isBuild = config.command === 'build'
+
+  const scan = config.command !== 'build' && config.optimizeDeps.devScan
 
   const sessionTimestamp = Date.now().toString()
 
@@ -100,7 +104,7 @@ export async function initDepsOptimizer(
   // If there wasn't a cache or it is outdated, we need to prepare a first run
   let firstRunCalled = !!cachedMetadata
   if (!cachedMetadata) {
-    if (isBuild) {
+    if (!scan) {
       // Initialize discovered deps with manually added optimizeDeps.include info
       const discovered = await initialProjectDependencies(
         config,
@@ -460,9 +464,14 @@ export async function initDepsOptimizer(
       exportsData: extractExportsData(resolved, config)
     })
 
-    // Debounced rerun, let other missing dependencies be discovered before
-    // the running next optimizeDeps
-    if (!isBuild) {
+    // Until the first optimize run is called, avoid triggering processing
+    // We'll wait until the user codebase is eagerly processed by Vite so
+    // we can get a list of every missing dependency before giving to the
+    // browser a dependency that may be outdated, thus avoiding full page reloads
+
+    if (scan || firstRunCalled) {
+      // Debounced rerun, let other missing dependencies be discovered before
+      // the running next optimizeDeps
       debouncedProcessing()
     }