+++ packages/vite/src/node/plugins/dynamicImportVars.ts
@@ -7,7 +7,12 @@ import { createFilter } from '@rollup/pluginutils'
 import { dynamicImportToGlob } from '@rollup/plugin-dynamic-import-vars'
 import type { Plugin } from '../plugin'
 import type { ResolvedConfig } from '../config'
-import { normalizePath, parseRequest, requestQuerySplitRE } from '../utils'
+import {
+  normalizePath,
+  parseRequest,
+  requestQuerySplitRE,
+  transformResult
+} from '../utils'
 
 export const dynamicImportHelperId = '/@vite/dynamic-import-helper'
 
@@ -204,10 +209,7 @@ export function dynamicImportVarsPlugin(config: ResolvedConfig): Plugin {
             `import __variableDynamicImportRuntimeHelper from "${dynamicImportHelperId}";`
           )
         }
-        return {
-          code: s.toString(),
-          map: config.build.sourcemap ? s.generateMap({ hires: true }) : null
-        }
+        return transformResult(s, importer, config)
       }
     }
   }