+++ playground/ssr-webworker/vite.config.js
@@ -1,7 +1,6 @@
-/**
- * @type {import('vite').UserConfig}
- */
-module.exports = {
+import { defineConfig } from 'vite'
+
+export default defineConfig({
   build: {
     minify: false
   },
@@ -32,4 +31,4 @@ module.exports = {
       }
     }
   ]
-}
+})